:rocket: rename functions and helpers
