+++ utils/canvas3Helpers.js
